Add unit class
This adds the unit class, there are few choices worth calling out:

- A lot of credit goes to Shopify/measured for the structure of this.
Some of the classes that will be named will be directly taken from
that, but by definition this gem is meant to convert things that really
can't be converted with any sort of guarantee.
- Currently the unit class has a lot of options. For now it is unclear
whether these options warrant different classes. For example, the
"approximate" flag will be used for things like "a dash" or "a pinch"
which do have unofficial conversions into tsp units. In those cases, the
approximate flag would only really be useful for signifying something in
a view or conversion that it's "more or less right."
- The "ingredient_unit" flag is a little different. For things like "a
package" or "a stick", those units are dependent on the ingredient that
is paired with them in order to determine if they are convertable or
comparable. That may genuinely be a different class, but for now a flag
seems appropriate until this fully takes shape.

Either way, these are built with the assumption that something like in
the file "scratch.rb" will be how this is used.